+++ api/src/main/java/gaiduchek/maksym/api/filters/ProductFilter.java
@@ -30,9 +30,9 @@ public class ProductFilter extends BaseFilter<Product> {
     private String name;
     private BigDecimal costMin;
     private BigDecimal costMax;
-    private List<Integer> companyIds;
-    private List<Integer> productCategoryIds;
-    private List<Integer> filterIds;
+    private List<Long> companyIds;
+    private List<Long> productCategoryIds;
+    private List<Long> filterIds;
     private List<Boolean> deleted;
 
     {
